fix(docs): split the grammar pin check into its own build-free audit
The grammar/core skew check guards highlighter correctness, not CSP, and it
only lived in the CSP script because that is where `packageJsonFor` was first
needed. Moving it out means `audit:csp` is back to one concern, and the pin
check runs in the lint job — no docs build in front of it, gated on the same
`packages` filter that already matches `pnpm-lock.yaml`.

Three fixes to the check itself while it moved:

- A dropped `@shikijs/langs` dependency reported `declares … as "undefined"`,
  which misnames the failure. It gets its own branch.
- The exact-pin regex rejected prerelease pins. Shiki publishes those and pins
  siblings exactly on them (`shiki@1.0.0-rc.0` -> `@shikijs/core: 1.0.0-rc.0`),
  so `EXACT_VERSION` now accepts prerelease and build metadata.
- The resolved version was read before the pin guards, so a broken install
  masked the pin diagnosis. It now resolves after both guards pass.

Closes #675
